requirements.txt
beautifulsoup4==4.12.2
lxml==5.1.0
pytest==7.4.4
python-dotenv==1.0.0
requests==2.31.0
